requirements.txt
pycaw
# pywinrt
psutil
pystray
Pillow
winrt-Windows.Foundation
winrt-Windows.Foundation.Collections
winrt-Windows.Media
winrt-Windows.Media.Audio
winrt-Windows.Media.Control
pywin32==306
PyYAML